Introduce CanvasCoursesContentWidget. Resuse on the list panel and in Academics. [#1714]
